:truck: api: move get-certification-details route and controller from lib to src
