+++ api/lib/application/certification-courses/certification-course-controller.js
@@ -3,17 +3,9 @@ import { usecases } from '../../domain/usecases/index.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as certifiedProfileRepository from '../../infrastructure/repositories/certified-profile-repository.js';
 import * as certificationCourseSerializer from '../../infrastructure/serializers/jsonapi/certification-course-serializer.js';
-import * as certificationDetailsSerializer from '../../infrastructure/serializers/jsonapi/certification-details-serializer.js';
 import * as certificationSerializer from '../../infrastructure/serializers/jsonapi/certification-serializer.js';
 import * as certifiedProfileSerializer from '../../infrastructure/serializers/jsonapi/certified-profile-serializer.js';
 import * as juryCertificationSerializer from '../../infrastructure/serializers/jsonapi/jury-certification-serializer.js';
-
-const getCertificationDetails = async function (request, h, dependencies = { certificationDetailsSerializer }) {
-  const certificationCourseId = request.params.id;
-  const certificationDetails = await usecases.getCertificationDetails({ certificationCourseId });
-
-  return dependencies.certificationDetailsSerializer.serialize(certificationDetails);
-};
 
 const getJuryCertification = async function (request, h, dependencies = { juryCertificationSerializer }) {
   const certificationCourseId = request.params.id;
@@ -88,7 +80,6 @@ const uncancel = async function (request, h) {
 };
 
 const certificationCourseController = {
-  getCertificationDetails,
   getJuryCertification,
   update,
   save,
