+++ api/src/certification/session-management/application/certification-details-route.js
@@ -1,0 +1,43 @@
+import Joi from 'joi';
+
+import { identifiersType } from '../../../../src/shared/domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { certificationDetailsController } from './certification-details-controller.js';
+
+const register = async function (server) {
+  server.route([
+    {
+      method: 'GET',
+      path: '/api/admin/certifications/{id}/details',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.hasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.certificationCourseId,
+          }),
+        },
+        handler: certificationDetailsController.getCertificationDetails,
+        tags: ['api'],
+        notes: [
+          'Cette route est utilisé par Pix Admin',
+          'Elle sert au cas où une certification a eu une erreur de calcul',
+          'Cette route ne renvoie pas le bon format.',
+        ],
+      },
+    },
+  ]);
+};
+
+const name = 'certification-details-api';
+export { name, register };
