+++ api/src/certification/session-management/routes.js
@@ -1,4 +1,5 @@
 import * as certificationCandidate from './application/certification-candidate-route.js';
+import * as certificationDetails from './application/certification-details-route.js';
 import * as certificationOfficer from './application/certification-officer-route.js';
 import * as certificationReport from './application/certification-report-route.js';
 import * as complementaryCertificationCourseResults from './application/complementary-certification-course-results-route.js';
@@ -20,6 +21,7 @@ const certificationSessionRoutes = [
   finalize,
   finalizedSession,
   certificationCandidate,
+  certificationDetails,
   invigilatorKit,
   juryComment,
   sessionForSupervising,
