+++ api/tests/unit/application/certification-courses/index_test.js
@@ -4,22 +4,6 @@ import { securityPreHandlers } from '../../../../src/shared/application/security
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | Certifications Course | Route', function () {
-  describe('GET /api/admin/certifications/{id}/details', function () {
-    it('should exist', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-      sinon.stub(certificationCoursesController, 'getCertificationDetails').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/admin/certifications/1234/details');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
-
   describe('GET /api/admin/certifications/{id}/certified-profile', function () {
     it('should exist', async function () {
       // given
